+++ sdk/data/azcosmos/cosmos_client_options.go
@@ -13,6 +13,8 @@ type ClientOptions struct {
 	// When EnableContentResponseOnWrite is false will cause the response to have a null resource. This reduces networking and CPU load by not sending the resource back over the network and serializing it on the client.
 	// The default is false.
 	EnableContentResponseOnWrite bool
+	// EnableSessionConsistency caches session tokens on writes and uses them for read operations.
+	EnableSessionConsistency bool
 	// PreferredRegions is a list of regions to be used when initializing the client in case the default region fails.
 	PreferredRegions []string
 }